audit A-items: harden the production path against silent failures
- select_float_type rejects non-float64 (kernels are double-only; cupy
  RawKernel would silently corrupt memory); config default float32->float64
- main.py: unhandled exception on any rank now comm.Abort()s the job
  instead of leaving peers blocked in a collective until walltime
- missing/typo'd config or [include] file raises FileNotFoundError instead
  of silently substituting parser defaults that differ from path_base.ini
- removed-legacy picture-path branch raises a clear error (was NameError)
- ROI strings corrected: snapshots are a 1 mm ROI, not 2 mm (log + docstring)
- --kill-on-bad-exit=1 added to the 5 launchers missing it
- README_Hermes.md -> legacy/ with a banner: it documents the legacy
  multi-level solver + a snapshot format the production pipeline does not
  write; live docs are README.md + experiments/README.md

Validated: negative tests for both new guards; plan_only + scan-path
smoke through the touched config/path stack.

+++ src/hermes/runtime/config.py
@@ -205,7 +205,7 @@ def load_config(path: str | Path) -> RunConfig:
 
     # simulation (optional)
     sim = cfg["simulation"] if "simulation" in cfg else {}
-    float_type_str = sim.get("float_type", "float32").strip()
+    float_type_str = sim.get("float_type", "float64").strip()
     tag = sim.get("tag", "run").strip()
 
     # laser (t_spot_on optional)
